Minor Feature : Reduced memory footprint in ram.
By storing only position and calculating the transform matrix inside the vertex shader, I store only a third of the otherwise chunky transformation matrix in the ram for each chunk.

ToDo: Reduce data needed to be stored in ram further by streaming or relying or bit shifting calculations and uniforms.